requirements.txt
fastapi
uvicorn[standard]
python-can
pydantic
pyusb
